Changed JS, HTML, and CSS signifcantly
Added IDs to the H1 and H3 elements for the heading text.  Moved all
the heading text into <header> and set header a bigger height.

Changed JS to use the IDs added in HTML which shortened the code by
quite a bit.

Added a moved around CSS rules to allow for changes in HTML and JS.