more minor html bugs
something happened with the last changes and padding worked the way it
was suppose to, which made an error in math I had made garlingly
obvious, where it hadn’t before.  Got rid of the padding and used
vertical-align properly.  (tried it before and it didn’t work, but I
used vetical-align: center, instead of vertical-align: middle);